voice table altered

+++ main.go
@@ -1697,7 +1697,7 @@ func queryServiceRequestReq(repos *repositoryServiceRequestReqArr, reqFrom int64
 	if err != nil {
 		return err
 	}
-	err = db.QueryRow(`update requests set reqnumber=luhn_generate($1) where id=$2 returning reqnumber`, id, id).Scan(&id)
+	err = db.QueryRow(`update requests set reqnumber=luhn_generate($1)::numeric where id=$2 returning reqnumber`, id, id).Scan(&id)
 	if err != nil {
 		return err
 	}
